src/requirements.txt
fastapi-0.124.4
uvicorn=0.33.0
python-multipart==0.0.20 # Gemini said It is required to support file uploads in your FastAPI application.
python-dotenv==1.0.1
pydantic-settings==2.8.1 # for configuration management and validation of settings in FastAPI applications.
 aiofiles-24.1.0 # to dead with files as a chunks